Ignore withdraw requests for users that don't exist

+++ services/dauth-service/src/services/backup/withdraw_backup.rs
@@ -14,19 +14,24 @@ pub async fn withdraw_backup(
     tracing::info!("Withdrawing backup");
 
     let mut transaction = context.local_context.database_pool.begin().await?;
-    let actual_network_id = database::backup_users::get(&mut transaction, user_id).await?;
+    let actual_network_id = database::backup_users::get(&mut transaction, user_id).await;
     transaction.commit().await?;
 
-    if actual_network_id != home_network_id {
-        Err(DauthError::InvalidMessageError(format!(
-            "Not the correct home network",
-        )))
+    if let Ok(actual_network_id) = actual_network_id {
+        if actual_network_id != home_network_id {
+            Err(DauthError::InvalidMessageError(format!(
+                "Not the correct home network",
+            )))
+        } else {
+            transaction = context.local_context.database_pool.begin().await?;
+            database::backup_users::remove(&mut transaction, user_id, home_network_id).await?;
+            database::auth_vectors::remove_all(&mut transaction, user_id).await?;
+            transaction.commit().await?;
+    
+            Ok(())
+        }
     } else {
-        transaction = context.local_context.database_pool.begin().await?;
-        database::backup_users::remove(&mut transaction, user_id, home_network_id).await?;
-        database::auth_vectors::remove_all(&mut transaction, user_id).await?;
-        transaction.commit().await?;
-
+        tracing::warn!(?user_id, "User is not being backed up by this network");
         Ok(())
     }
 }
